Move JUnitXml parsing from RunnerBase to NoseRunner
